+++ packages/eds-core-react/src/components/Select/NativeSelect.test.tsx
