+++ packages/eds-core-react/src/components/Select/NativeSelect.tokens.ts
