--- packages/eds-core-react/src/components/Select/NativeSelect/index.ts
@@ -1,1 +1,0 @@
-export * from './NativeSelect'
